Rename directory to position_controller
