+++ src/software/embedded/position_controller/pid_controller_test.cpp
@@ -1,4 +1,4 @@
-#include "software/embedded/velocity_controller/pid_controller.h"
+#include "software/embedded/position_controller/pid_controller.h"
 
 #include <gtest/gtest.h>
